+++ .github/workflows/python-app.yml
@@ -23,8 +23,6 @@ jobs:
       uses: actions/setup-python@v3
       with:
         python-version: "3.10"
-    - name: Install system dependencies
-      run: sudo apt-get update && sudo apt-get install -y libpq-dev gcc
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
