--- app/main.py
@@ -1,8 +1,0 @@
-from api import API
-
-def main():
-    api = API()
-    api.run()
-
-if __name__ == "__main__":
-    main()
