+++ config.py
@@ -1,5 +1,3 @@
-import os
-
 import requests
 
 API_HOST_URL = '127.0.0.1'
@@ -9,7 +7,6 @@ STEAMAPI_BASE_URL = 'https://api.steampowered.com/'
 STEAMSTORE_BASE_URL = 'https://store.steampowered.com/api/'
 
 APPS_LIST_CACHE_FILE = 'cache/apps_list.json'
-ADDED_GAMES_LIST_CACHE_FILE = 'cache/added_games_list.txt'
 CACHE_EXPIRATION_TIME = 604800  # Time in seconds (604800 seconds = 1 week)
 
 def fetch_from_api(endpoint):
@@ -87,25 +84,6 @@ def set_womp_exception(enabled):
         sys.excepthook = womp_exception_hook
     else:
         sys.excepthook = sys.__excepthook__
-
-def load_env():
-    """Load environment variables from the .env file."""
-    with open(".env") as f:
-        for line in f:
-            if line.strip() and not line.startswith("#"):  # Ignore empty lines and comments
-                key, value = line.strip().split("=", 1)
-                os.environ[key] = value
-                # update DB_CONFIG
-                if key in DB_CONFIG:
-                    DB_CONFIG[key] = value
-
-DB_CONFIG = {
-    "DB_USER": None,
-    "DB_PASSWORD": None,
-    "DB_NAME": None,
-    "DB_HOST": None,
-    "DB_PORT": None,
-}
 
 # Set the custom exception hook
 set_womp_exception(True)
