+++ main.py
@@ -1,0 +1,2 @@
+def hello():
+    print("Hello world")
