+++ requirements.txt
@@ -1,5 +1,3 @@
 uvicorn~=0.34.0
 fastapi~=0.115.7
-SQLAlchemy~=2.0.37
 requests~=2.32.3
-psycopg2~=23.2.1
